First Reimplementation of the Table_Val
First Reimplementation of the Table_Validation to check if the examples are also in the same rows aswell as the Querying Alg for the corresponding y Values of x in Q, but all just for single-col as of right now. Could not be tested yet as Eduroam and Cisco say No!